Create a new file, only being used for editing posts
The Posts.php page has been linked to the necessary pages for editing and deleting posts. The editpost page is a variation of the addnewpost page. It shows the currently submitted information for that page. However, this will not update the db yet, that functionality has yet to be created.